GEOMESA-3033 Provide metrics reporting for Kafka data store cache
* New geomesa-metrics modules ported from geomesa-convert
