+++ docs/user/kafka/usage.rst
@@ -71,6 +71,8 @@ Parameter                            Type    Description
                                              See :ref:`kafka_ssi`
 ``kafka.serialization.lazy``         Boolean Use lazy deserialization of features. This may improve processing load at
                                              the expense of slightly slower query times
+``kafka.metrics.reporters``          String  Reporters used to publish Kafka metrics, as TypeSafe config. To use multiple reporters, nest
+                                             them under the key ``reporters``. See :ref:`geomesa_metrics` for details
 ``geomesa.query.loose-bounding-box`` Boolean Use loose bounding boxes, which offer improved performance but are not exact
 ``geomesa.query.audit``              Boolean Audit incoming queries. By default audits are written to a log file
 ``geomesa.security.auths``           String  Default authorizations used to query data, comma-separated
