+++ geomesa-convert/geomesa-convert-metrics-ganglia/pom.xml
@@ -29,23 +29,8 @@
             <artifactId>geomesa-convert-common_${scala.binary.version}</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.dropwizard.metrics</groupId>
-            <artifactId>metrics-ganglia</artifactId>
-            <version>${metrics.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.specs2</groupId>
-            <artifactId>specs2-core_${scala.binary.version}</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.specs2</groupId>
-            <artifactId>specs2-junit_${scala.binary.version}</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-log4j12</artifactId>
-            <scope>test</scope>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-metrics-ganglia_${scala.binary.version}</artifactId>
         </dependency>
     </dependencies>
