+++ geomesa-convert/geomesa-convert-metrics-ganglia/src/main/scala/org/locationtech/geomesa/ScalaDocs.scala
@@ -1,0 +1,16 @@
+/***********************************************************************
+ * Copyright (c) 2013-2021 Commonwealth Computer Research, Inc.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Apache License, Version 2.0
+ * which accompanies this distribution and is available at
+ * http://www.opensource.org/licenses/apache2.0.php.
+ ***********************************************************************/
+
+package org.locationtech.geomesa
+
+/**
+  * Placeholder class to force generation of -javadocs and -sources jars for maven publishing
+  */
+object ScalaDocs {
+
+}
