+++ geomesa-convert/geomesa-convert-metrics-graphite/pom.xml
@@ -17,23 +17,8 @@
             <artifactId>geomesa-convert-common_${scala.binary.version}</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.dropwizard.metrics</groupId>
-            <artifactId>metrics-graphite</artifactId>
-            <version>${metrics.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.specs2</groupId>
-            <artifactId>specs2-core_${scala.binary.version}</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.specs2</groupId>
-            <artifactId>specs2-junit_${scala.binary.version}</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-log4j12</artifactId>
-            <scope>test</scope>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-metrics-graphite_${scala.binary.version}</artifactId>
         </dependency>
     </dependencies>
