+++ geomesa-kafka/geomesa-kafka-datastore/src/main/scala/org/locationtech/geomesa/kafka/data/KafkaDataStoreFactory.scala
@@ -21,6 +21,7 @@ import org.locationtech.geomesa.index.metadata.MetadataStringSerializer
 import org.locationtech.geomesa.kafka.data.KafkaDataStore._
 import org.locationtech.geomesa.kafka.utils.GeoMessageSerializer.GeoMessageSerializerFactory
 import org.locationtech.geomesa.memory.cqengine.utils.CQIndexType
+import org.locationtech.geomesa.metrics.core.GeoMesaMetrics
 import org.locationtech.geomesa.security
 import org.locationtech.geomesa.security.AuthorizationsProvider
 import org.locationtech.geomesa.utils.audit.{AuditLogger, AuditProvider, NoOpAuditProvider}
@@ -191,6 +192,13 @@ object KafkaDataStoreFactory extends GeoMesaDataStoreInfo with LazyLogging {
     }
     val authProvider = buildAuthProvider(params)
 
+    val metrics = MetricsReporters.lookupOpt(params).map { conf =>
+      val config = ConfigFactory.parseString(conf).resolve()
+      val reporters =
+        if (config.hasPath("reporters")) { config.getConfigList("reporters").asScala } else { Seq(config) }
+      GeoMesaMetrics(catalog, reporters)
+    }
+
     val ns = Option(NamespaceParam.lookUp(params).asInstanceOf[String])
 
     // noinspection ScalaDeprecation
@@ -201,7 +209,7 @@ object KafkaDataStoreFactory extends GeoMesaDataStoreInfo with LazyLogging {
     }
 
     KafkaDataStoreConfig(catalog, brokers, zookeepers, consumers, producers, clearOnStart, topics, serialization,
-      indices, looseBBox, authProvider, audit, ns)
+      indices, looseBBox, authProvider, audit, metrics, ns)
   }
 
   private def buildAuthProvider(params: java.util.Map[String, Serializable]): AuthorizationsProvider = {
