+++ geomesa-metrics/geomesa-metrics-cloudwatch/src/main/resources/META-INF/services/org.locationtech.geomesa.metrics.core.ReporterFactory
