+++ geomesa-metrics/geomesa-metrics-core/src/main/resources/META-INF/services/org.locationtech.geomesa.metrics.core.ReporterFactory
@@ -1,0 +1,2 @@
+org.locationtech.geomesa.metrics.core.ConsoleReporterFactory
+org.locationtech.geomesa.metrics.core.Slf4jReporterFactory
