+++ geomesa-metrics/geomesa-metrics-ganglia/src/main/resources/META-INF/services/org.locationtech.geomesa.metrics.core.ReporterFactory
