+++ geomesa-metrics/pom.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <parent>
+        <artifactId>geomesa_2.11</artifactId>
+        <groupId>org.locationtech.geomesa</groupId>
+        <version>3.2.0-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>geomesa-metrics_2.11</artifactId>
+    <name>GeoMesa Metrics Parent</name>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>geomesa-metrics-core</module>
+        <module>geomesa-metrics-cloudwatch</module>
+        <module>geomesa-metrics-ganglia</module>
+        <module>geomesa-metrics-graphite</module>
+    </modules>
+
+</project>
